+++ src/features/authentication/password-forgotten/model/email-step/use-email-step.test.ts
@@ -1,9 +1,9 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
-import { signInWithOneTimePassword } from '../sign-in-with-otp'
+import { signInWithOneTimePassword } from '../../service/sign-in-with-otp'
 import { useEmailStep } from './use-email-step'
 
-vi.mock('../sign-in-with-otp', () => ({
+vi.mock('../../service/sign-in-with-otp', () => ({
   signInWithOneTimePassword: vi.fn()
 }))
