+++ src/features/authentication/password-forgotten/model/email-step/use-email-step.ts
@@ -1,6 +1,6 @@
 import { ref } from 'vue'
 
-import { signInWithOneTimePassword } from '../sign-in-with-otp'
+import { signInWithOneTimePassword } from '../../service/sign-in-with-otp'
 
 /**
  * Handles OTP mail request for the email step.
