+++ src/features/authentication/password-forgotten/model/otp-step/use-otp-step.test.ts
@@ -1,9 +1,9 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
-import { checkOneTimePasswordByRecovery } from './check-otp-by-recovery'
+import { checkOneTimePasswordByRecovery } from '../../service/otp-step/check-otp-by-recovery'
 import { useOtpStep } from './use-otp-step'
 
-vi.mock('./check-otp-by-recovery')
+vi.mock('../../service/otp-step/check-otp-by-recovery')
 
 describe('useOtpStep', () => {
   beforeEach(() => {
