+++ src/features/settings/index.ts
@@ -1,8 +1,8 @@
 export { default as settingsDe } from './i18n/de'
 export { default as settingsEn } from './i18n/en'
 export { default as translationKeys } from './i18n/keys'
-export { getInitialLanguage } from './model/set-language/language-service'
-export { getInitialTheme } from './model/set-theme/theme-service'
-export { getThemeFromStorage, setThemeToStorage } from './model/set-theme/theme-storage'
+export { getInitialLanguage } from './service/set-language/language-service'
+export { getInitialTheme } from './service/set-theme/theme-service'
+export { getThemeFromStorage, setThemeToStorage } from './service/set-theme/theme-storage'
 export { default as LanguageSwitch } from './ui/LanguageSwitch.vue'
 export { default as ThemeToggle } from './ui/ThemeToggle.vue'
